Make Quote not have pointers
Originally I made the Quote type use pointers so that we could tell the
difference between no-value-present (nil) and actually-zero-value.
That's fine but the client gets exposed to it.

In the new design, the backend converts from the pointer based structure
to a value-based structure. The client never need see the quote-based
struct. The value -1 is used for Bid / Ask / Last not present in the
quote.

Also, the Market type has been changed to just store a quote.